refactor: :recycle: clean up and enhance documentation across multiple files for clarity and consistency
