+++ NET10/Demo.Application/Commands/ICommandJob.cs
@@ -1,17 +1,11 @@
-namespace GamePulse.Commands;
+namespace Demo.Application.Commands;
 
-/// <summary>
-/// 
-/// </summary>
 public interface ICommandJob
 {
     /// <summary>
-    /// 
+    /// Executes the command asynchronously using the provided service provider and supports cancellation.
     /// </summary>
     /// <param name="serviceProvider"></param>
     /// <param name="ct"></param>
-    /// <summary>
-/// Executes the command asynchronously using the provided service provider and supports cancellation.
-/// </summary>
     Task ExecuteAsync(IServiceProvider serviceProvider, CancellationToken ct);
 }
