+++ NET10/Demo.Application/Demo.Application.csproj
@@ -23,8 +23,4 @@
       <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.4" />
     </ItemGroup>
 
-    <ItemGroup>
-      <ProjectReference Include="..\GamePulse\GamePulse.csproj" />
-    </ItemGroup>
-
 </Project>
