+++ NET10/Demo.Application/Services/IpToNationService.cs
@@ -1,22 +1,23 @@
-﻿using GamePulse.Repositories;
-using GamePulse.Repositories.IpToNation.Cache;
+﻿using Demo.Application.Repositories;
 
-namespace GamePulse.Services.IpToNation;
+namespace Demo.Application.Services;
 
 public class IpToNationService : IIpToNationService
 {
     private readonly IIpToNationCache _cache;
     private readonly IIpToNationRepository _repo;
+    private readonly ITelemetryService _telemetryService;
 
-    public IpToNationService(IIpToNationCache cache, IIpToNationRepository repo)
+    public IpToNationService(IIpToNationCache cache, IIpToNationRepository repo, ITelemetryService telemetryService)
     {
         _cache = cache;
         _repo = repo;
+        _telemetryService = telemetryService;
     }
 
     public async Task<string> GetNationCodeAsync(string clientIp, CancellationToken ct)
     {
-        using var span = GamePulseActivitySource.StartActivity("GetNationCodeAsync");
+        using var span = _telemetryService.StartActivity("GetNationCodeAsync");
         /*
         var result = await _cache.GetAsync(clientIp);
         if (result.IsSuccess)
