+++ NET10/Demo.Infra/Demo.Infra.csproj
@@ -14,8 +14,11 @@
       <PackageReference Include="Dapper" Version="2.1.66" />
       <PackageReference Include="Dapper.Extensions.PostgreSQL" Version="5.3.1" />
       <PackageReference Include="FluentResults" Version="4.0.0" />
+      <PackageReference Include="IP2Location.IPGeolocation" Version="9.0.3" />
       <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="2.3.0" />
       <PackageReference Include="Npgsql.OpenTelemetry" Version="9.0.2" />
+      <PackageReference Include="OpenTelemetry.Instrumentation.StackExchangeRedis" Version="1.12.0-beta.2" />
+      <PackageReference Include="StackExchange.Redis" Version="2.8.58" />
     </ItemGroup>
 
 </Project>
