+++ NET10/Demo.Web/Demo.Web.csproj
@@ -16,6 +16,7 @@
         <PackageReference Include="LiteBus.Messaging" Version="3.0.0" />
         <PackageReference Include="LiteBus.Queries" Version="3.0.0" />
         <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.7" />
+        <PackageReference Include="Scalar.AspNetCore" Version="2.6.5" />
         <PackageReference Include="Serilog" Version="4.3.0" />
         <PackageReference Include="Serilog.AspNetCore" Version="9.0.0" />
         <PackageReference Include="Serilog.Extensions.Logging" Version="9.0.2" />
