+++ NET10/GamePulse/EndPoints/Login/LoginEndpointV1.cs
@@ -9,9 +9,6 @@ using GamePulse.Services.Auth;
 
 namespace GamePulse.EndPoints.Login;
 
-/// <summary>
-/// 
-/// </summary>
 public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
 {
     private readonly IAuthService _authService;
@@ -19,22 +16,17 @@ public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
     private readonly Tracer _tracer;
 
     /// <summary>
-    /// 
+    /// Initializes a new instance of the <see cref="LoginEndpointV1"/> class with the specified authentication service, logger, and tracer.
     /// </summary>
     /// <param name="authService"></param>
     /// <param name="logger"></param>
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LoginEndpointV1"/> class with the specified authentication service, logger, and tracer.
-    /// </summary>
     public LoginEndpointV1(IAuthService authService, ILogger<LoginEndpointV1> logger, Tracer tracer)
     {
         _logger = logger;
         _authService = authService;
         _tracer = tracer;
     }
-    
-    /// <summary>
-    /// 
+
     /// <summary>
     /// Configures the login endpoint with versioning, route, anonymous access, validation error logging, rate limiting, and API documentation summary.
     /// </summary>
@@ -43,7 +35,7 @@ public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
         Version(1);
         Post("/api/login");
         AllowAnonymous();
-        
+
         PreProcessor<ValidationErrorLogger<LoginRequest>>();
         // 헤더 이름은 원하는 대로 설정할 수 있습니다.
         // 지정하지 않으면 라이브러리는 들어오는 요청에서 X-Forwarded-For 헤더의 값을 읽으려고 시도합니다.
@@ -53,7 +45,7 @@ public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
         Throttle(
             hitLimit: 60,
             durationSeconds: 60);
-        Summary(s => 
+        Summary(s =>
         {
             s.Summary = "사용자 로그인";
             s.Description = "입력한 정보를 바탕으로 로그인을 수행합니다.";
@@ -62,7 +54,7 @@ public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
     }
 
     /// <summary>
-    /// 
+    ///
     /// </summary>
     /// <param name="req"></param>
     /// <param name="ct"></param>
@@ -70,12 +62,12 @@ public class LoginEndpointV1 : Endpoint<LoginRequest, TokenResponse>
     {
         using var span = _tracer.StartActiveSpan("Login");
         _logger.LogInformation("{@Request}", req);
-        
+
         if (await _authService.CredentialsAreValidAsync(req.Username, req.Password, ct) == false)
         {
             ThrowError("Invalid username and password");
         }
-        
+
         var faker = new Faker();
         var userId = faker.Random.Uuid().ToString();
         Response = await CreateTokenWith<MyTokenService>(userId, u =>
