+++ NET10/GamePulse/GamePulse.csproj
@@ -41,7 +41,11 @@
   <ItemGroup>
     <Folder Include="k6\" />
     <Folder Include="logs\" />
-    <Folder Include="Utils\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Demo.Application\Demo.Application.csproj" />
+    <ProjectReference Include="..\Demo.Infra\Demo.Infra.csproj" />
   </ItemGroup>
 
 </Project>
