+++ NET10/GamePulse/Processors/ValidationErrorLogger.cs
@@ -2,16 +2,12 @@ using FastEndpoints;
 
 namespace GamePulse.Processors;
 
-/// <summary>
-/// 
-/// </summary>
-/// <typeparam name="TRequest"></typeparam>
 public class ValidationErrorLogger<TRequest> : IPreProcessor<TRequest>
 {
     private readonly ILogger<ValidationErrorLogger<TRequest>> _logger;
 
     /// <summary>
-    /// 
+    ///
     /// </summary>
     /// <summary>
     /// Initializes a new instance of the <see cref="ValidationErrorLogger{TRequest}"/> class.
@@ -22,11 +18,6 @@ public class ValidationErrorLogger<TRequest> : IPreProcessor<TRequest>
     }
 
     /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="context"></param>
-    /// <param name="ct"></param>
-    /// <summary>
     /// Logs validation errors for the given request if any are present in the context.
     /// </summary>
     /// <param name="context">The pre-processor context containing the request and validation results.</param>
@@ -36,7 +27,7 @@ public class ValidationErrorLogger<TRequest> : IPreProcessor<TRequest>
     {
         if (context.ValidationFailures.Count > 0)
         {
-            _logger.LogWarning("Validation failed for {RequestType}: {Errors}", 
+            _logger.LogWarning("Validation failed for {RequestType}: {Errors}",
                 typeof(TRequest).Name,
                 string.Join(", ", context.ValidationFailures.Select(f => $"{f.PropertyName}: {f.ErrorMessage}")));
         }
